Upgrade to GPLv3
The current GPLv2 license includes this text:

  This program is free software; you can redistribute it and/or
  modify it under the terms of the GNU General Public License as
  published by the Free Software Foundation; either version 2, or (at
  your option) any later version.

which clearly allows for the license to be rolled forward to a later
GPL version.  All past contributions were made with this license in
place, so I think we can assume contributors are OK with the license
being rolled forward.  I see no reason why we shouldn't move to GPLv3,
so lets do that.